+++ odoo-openmrs/src/main/java/com/ozonehis/eip/odoo/openmrs/processors/MedicationRequestProcessor.java
@@ -85,7 +85,7 @@ public class MedicationRequestProcessor implements Processor {
 
                 Integer companyId = null;
                 if (enableEncounterLocationCompanyMapping) {
-                    companyId = getCompanyIdByEncounterLocationUuid(encounter);
+                    companyId = companyHandler.getCompanyIdByEncounterLocation(encounter);
                     if (companyId == null) {
                         log.warn(
                                 "Skipping MedicationRequest sync as company id is null for MedicationRequest id {}",
@@ -136,37 +136,5 @@ public class MedicationRequestProcessor implements Processor {
         } catch (Exception e) {
             throw new CamelExecutionException("Error processing MedicationRequest", exchange, e);
         }
-    }
-
-    private Integer getCompanyIdByEncounterLocationUuid(Encounter encounter) {
-        String locationUuid = resolveEncounterLocationUuid(encounter);
-        if (locationUuid == null) {
-            log.warn("Encounter {} has no location reference", encounter.getIdPart());
-            return null;
-        }
-        Integer companyId = companyHandler.getCompanyIdByExternalId(locationUuid);
-        if (companyId == null) {
-            log.warn(
-                    "No res.company external id matches location uuid {} for encounter {}",
-                    locationUuid,
-                    encounter.getIdPart());
-            return null;
-        }
-        return companyId;
-    }
-
-    private String resolveEncounterLocationUuid(Encounter encounter) {
-        if (encounter == null || !encounter.hasLocation()) {
-            return null;
-        }
-        String uuid = encounter
-                .getLocationFirstRep()
-                .getLocation()
-                .getReferenceElement()
-                .getIdPart();
-        if (uuid == null || uuid.isBlank()) {
-            return null;
-        }
-        return uuid.trim();
     }
 }
